requirements.txt
tensorflow==2.1.0
grpcio==1.24.3
opencv-python==4.1.0.25
pandas
Pillow
matplotlib
numpy